+++ homebrew/Casks/apidae.rb
@@ -1,5 +1,5 @@
 cask "apidae" do
-  version "0.1.0"
+  version "0.2.0"
   sha256 :no_check
 
   url "https://github.com/MrADeveci/getapidae/releases/download/v#{version}/Apidae.dmg"